:truck: split tests into modules
